refactor: reorganize project structure under app
